% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{changecolnamesInRules}
\alias{changecolnamesInRules}
\title{-----------------------------------------------------------------------------}
\usage{
changecolnamesInRules(dataset, rules)
}
\arguments{
\item{dataset}{the dataframe on which rules are to be run}

\item{rules}{rules defined in a csv file 
-----------------------------------------------------------------------------}
}
\value{
rules in required format
}
\description{
: This function is used to remove the . or _ in column names of
 the dataset present in the rules file
-----------------------------------------------------------------------------
}
