% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{formatOutput}
\alias{formatOutput}
\title{Format output into required format}
\usage{
formatOutput(dataset, outputDf, rules, filteredDataFalse, input.columns,
  ruleNum)
}
\arguments{
\item{dataset}{rules defined in a csv file}

\item{outputDf}{the output data frame returned by the executeRulesOnDataset 
function}

\item{rules}{the rules defined in csv format}

\item{filteredDataFalse}{the fitlered data which is flagged as false}

\item{input.columns}{the input columns of the dataset which is obtained
from getrequiredColumns function}

\item{ruleNum}{the number of the rule currently being run}
}
\value{
output in required format
}
\description{
: This function is used to change the output of the rules 
involving groupby
}
\keyword{internal}
