% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{getConditionForMultiGroupBy}
\alias{getConditionForMultiGroupBy}
\title{-----------------------------------------------------------------------------}
\usage{
getConditionForMultiGroupBy(groupByColumn, aggregationFunc, aggregationColumn)
}
\arguments{
\item{groupByColumn}{columns to groupby}

\item{aggregationFunc}{aggregation function to be applied}

\item{aggregationColumn}{column on which aggregation function is to be 
applied  
-----------------------------------------------------------------------------}
}
\value{
required rule condition
}
\description{
: This function is used to get the required rule condition when 
there are multiple groupby columns
-----------------------------------------------------------------------------
}
