% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{getDrlForRowwiseRules}
\alias{getDrlForRowwiseRules}
\title{-----------------------------------------------------------------------------}
\usage{
getDrlForRowwiseRules(dataset, rules, ruleNum, inputColumns, outputColumns)
}
\arguments{
\item{dataset}{rules defined in a csv file}

\item{rules}{the rules defined in csv format
-----------------------------------------------------------------------------}

\item{outputDf}{the output dataframe returned by the executeRulesOnDataset 
function}
}
\value{
drl format of rules for row wise rules
}
\description{
: This function is used to get the drl format of rules for row 
wise rules
-----------------------------------------------------------------------------
}
