% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{getrequiredColumns}
\alias{getrequiredColumns}
\title{-----------------------------------------------------------------------------}
\usage{
getrequiredColumns(dataset, rules)
}
\arguments{
\item{dataset}{the data frame on which rules are to be run}

\item{rules}{rules defined in a csv file
-----------------------------------------------------------------------------}
}
\value{
required input columns and output columns
}
\description{
: This function is used to get the required input and output 
columns
-----------------------------------------------------------------------------
}
\keyword{internal}
