% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools_package.R
\docType{package}
\name{Rdrools}
\alias{Rdrools}
\alias{Rdrools-package}
\title{Rdrools}
\description{
An interface for using the popular Java based Drools, which is a Business Rule Management System (See <https://www.drools.org> for more information). 
 This package provides data scientists an intuitive interface to execute business rules on datasets for the purpose of analysis or designing intelligent systems, while leveraging the Drools rule engine. 
 Rules can be defined in a simple format in a CSV, imported into the R Session as a data frame, and passed directly to the rule-checking functions.
 Rules written in DRL format, accepted natively by Drools, can also be executed through an R interface. Credits to Mu Sigma for their continued support throughout the development of the package.
}
