% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{ruleToCompareColumns}
\alias{ruleToCompareColumns}
\title{-----------------------------------------------------------------------------}
\usage{
ruleToCompareColumns(dataset, rules, ruleNum)
}
\arguments{
\item{dataset}{data frame on which rules are to be run}

\item{rules}{the rules defined in csv format}

\item{ruleNum}{number of the rule which has condition to compare columns
-----------------------------------------------------------------------------}
}
\value{
output in required format
}
\description{
: This function is used to get drl rule for rules involving
 comparing columns
-----------------------------------------------------------------------------
}
