% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{runRulesDrl}
\alias{runRulesDrl}
\title{Apply a set of rule transformations to a data frame}
\usage{
runRulesDrl(rulesSession, inputDf)
}
\arguments{
\item{rulesSession}{a session of the rules engine created using the the
\code{\link {rulesSessionDrl}}}

\item{inputDf}{a data frame consisting of a set of rows you wish to transform, and columns you wish to use in the transformation}
}
\value{
outputDf a data frame which is the result of transformations applied to the input data frame(\code{inputDf}), the columns being the list provided through the \code{outputColumns} parameter in 
\code{\link {rulesSessionDrl}}
}
\description{
This function is the core of the Rdrools package. Rules are applied on an input data frame and the results are returned as the output of the function. The columns on which the rules need to be applied have to be provided explicitly. Additionally, the new columns that would be created based on the rules have to be provided explicitly as well.\cr 
The rules engine picks up a row from the data frame, applies the transformation to it based on rules provided and saves the result in an output data frame.
}
\details{
If you are not familiar with the drools file format, please have a look at the references provided in the \code{\link {Rdrools man page}}. More details on how conflicting rules are resolved using either salience or the Reete algorithm are also present in the references.
}
\examples{
library(Rdrools)
data(class)
data(rules)
inputColumns<-c('name', 'class', 'grade', 'email')
outputColumns<-c('address', 'subject', 'body')
rulesSession<-rulesSession(rules, inputColumns, outputColumns)
outputDf<-runRules(rulesSession, class)


}
\seealso{
\code{\link{runRulesDrl}}, \code{\link{Rdrools}}'
}
\author{
Ashwin Raaghav <ashraaghav@gmail.com>, SMS Chauhan <smschauhah@gmail.com>
}
