% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactionData}
\alias{transactionData}
\title{Sample transaction data}
\format{A data frame with 39999 rows and 16 variables:}
\usage{
transactionData
}
\description{
A dataset containing banking transactions, with variables such as number of transactions, transaction value,
and so on
}
\keyword{datasets}
