% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactionRules}
\alias{transactionRules}
\title{Sample rules for the sample transaction dataset (transactionData)}
\format{A data frame with 7 rows and 6 variables:
\describe{
  \item{Filters}{Filters to be applied in a specific rule}
  \item{GroupBy}{Grouping conditions to be appled in a specific rule}
  \item{Column}{The variable on which the function is to applied in a specific rule}
  \item{Function}{Function to be applied on the specific variable in a rule}
  \item{Operation}{The comparison operation to be performed to check the rule}
  \item{Argument}{The value agains which the comparison operation is to be performed}
  }}
\usage{
transactionRules
}
\description{
A dataset containing the sample rules to be applied on the transaction dataset, where each
row represents a rule
}
\keyword{datasets}
