% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class}
\alias{class}
\title{Sample data of students' grades}
\format{A data frame with 15 observations and 5 variables
\describe{
  \item{id}{ID of the student}
  \item{name}{Name of the student}
  \item{class}{Subject for which the grade is specified}
  \item{grade}{Grade in the subject}
  \item{email}{E-mail ID of the student}
  }}
\usage{
class
}
\description{
A dataset containing students' grades in different subjects
}
\keyword{datasets}
