% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools.R
\name{getDrlForFilterRules}
\alias{getDrlForFilterRules}
\title{-----------------------------------------------------------------------------}
\usage{
getDrlForFilterRules(dataset, rules, ruleNum, outputCols, input.columns,
  output.columns)
}
\arguments{
\item{dataset}{data frame on which rules are to be run}

\item{rules}{rules in csv format}

\item{ruleNum}{the number of the current rule}

\item{outputCols}{the output statments to show the output}

\item{input.columns}{input columns which are obtained from getrequiredColumns}

\item{output.columns}{output columns which are obtained from 
getrequiredColumns
-----------------------------------------------------------------------------}
}
\value{
filtered output with flags true/false
}
\description{
: This function is used to get the drl format for the rules 
which have filters
-----------------------------------------------------------------------------
}
\keyword{internal}
