% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrools_package.R
\docType{package}
\name{Rdrools}
\alias{Rdrools}
\alias{Rdrools-package}
\title{Rdrools}
\description{
An interface for using the popular Java based Drools, which is a Business Rule Management System (See <https://www.drools.org> for more information).
This package allows data scientists to run rules on datasets provided through a CSV file using common data science verbs such as Filter, Group By and Aggregate.
Additionally, rules defined directly in the DRL format accepted by the Drools engine can also be run.
Credits to Mu Sigma for their continued support throughout the development of the package.
}
