% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rules}
\alias{rules}
\title{Sample rules in DRL format}
\format{An object of class \code{character} of length 26.}
\usage{
rules
}
\description{
A list of character strings, providing an example for rules in the DRL format accepted by the 
Drools engine
}
\keyword{datasets}
